# login-prediction-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Login Prediction API",
        "description": "Kullanıcıların login tahminleri için API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Tüm kullanıcıları listeler",
                "operationId": "aa95fa19b1a57989cefc23ddba466b66",
                "responses": {
                    "200": {
                        "description": "Başarılı işlem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/usernames": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Tüm kullanıcı adlarını listeler",
                "operationId": "b5e8222f4dc318135384e14dd5915634",
                "responses": {
                    "200": {
                        "description": "Başarılı işlem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{userId}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Belirli bir kullanıcıyı getirir",
                "operationId": "87a1049ade8810e518928db4e12ea5b6",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Kullanıcı ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı işlem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kullanıcı bulunamadı"
                    }
                }
            }
        },
        "/api/users/{userId}/predictions": {
            "get": {
                "tags": [
                    "Predictions"
                ],
                "summary": "Kullanıcı için login tahminlerini getirir",
                "operationId": "72d05ca3dbd0262ed5199579b2ebf32d",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Kullanıcı ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı işlem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userId": {
                                            "type": "string"
                                        },
                                        "predictions": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string",
                                                "format": "date-time",
                                                "nullable": true
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kullanıcı bulunamadı"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "logins": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "Users"
        },
        {
            "name": "Predictions",
            "description": "Predictions"
        }
    ]
}